Use non-moving nursery collections for StickyImmix
This is due to the fact that Android registers weak references when
scanning objects instead of when creating the weak references. If MMTk
moves a live object during a nursery GC, then any pointer from a mature
object (in the reference processing queue) to the live nursery object
breaks. This will later cause the ReferenceQueueDaemon thread to SIGSEGV
when it is processing the weak references.

Disabling movement for the nursery collections is a stop-gap measure and
we will revisit this in the future.