Fix clippy warnings for Rust 1.79 (#1151)
This PR fixes new warnings generated by Clippy after upgrading Rust to
1.79.

- Only define `ImmixSpaceArg::mixed_age` when the "vo_bit" feature is
enabled. This eliminates a dead code warning.
- Use the associated constant `<integer_type>::MAX` (and `MIN`, too)
instead of the deprecated `max_value()` method and the
`::std::<integer_type>::MAX` constant.
-   Annotate the destination type of some `transmute` calls.
- Use the safe `std::ptr::NonNull::from()` converter instead of
`transmute`.
-   Initialize array elements of type `Atomic<MapState>` from constant.
- Remove `MutatorContext::barrier_impl`. It is not used by any bindings
right now, and it depends on the layout of `&dyn Trait`. Since `Barrier`
has `Downcast` as a supertrait, if a binding needs a reference to a
concrete `Barrier` implementation, it should just use downcast.

The warning against the `transmute` call in `mock_vm.rs` is suppressed
because it is almost impossible to annotate.

Also fixes compilation problems:

- Made benchmarks related to the "mock_test" feature conditionally
compiled so that we don't get compilation error when running `cargo
clippy --benches` or `cargo clippy --all-targets`.