Clear pin bits (#1166)
If an object dies, we need to make sure his pin bit gets cleared. This
PR enables clearing the pin bits when sweeping - clearing the whole
block when `super::BLOCK_ONLY` is true, or at line granularity when a
line is not marked.
I've only added it to the immix policy, since for everything else
pinning is either unsupported or it's a NOP.