Add support for Zygote space in MMTk
We add a `ZygoteSpace` to `CommonPlan` which is a wrapper around an
`ImmixSpace`. This allows us to allocate non-moving objects as well as
moving objects into the same space. We reuse the same `ImmixAllocator`
by rebinding its internal `ImmixSpace` to the actual `ImmixSpace` after
the first Zygote fork.

We only do full-heap GCs for the Zygote process before the Zygote space
has been created to simplify the implementation as well as maximally
compact the Zygote space.

We remove the `LOCAL_LOS_MARK_NURSERY` metadata as the cycling of the
bits was causing issues when visiting objects to generate the bootimage.
If the mark state was `0`, uninitialized metadata for an arbitrary
address would result in the heap visitor thinking the address is an
object.

We also remove the `MarkSweepSpace`-based non-moving space and replace
it with an `ImmortalSpace`. We now allocate non-moving objects directly
into the `ImmixSpace` and set the pin-bits. This also allows us
maximally compact the `ZygoteSpace` before the first Zygote fork.

For 32-bit builds, we always force a full-heap GC since we currently do
not have the write barriers implemented for them.

This commit allows us to boot the Cuttlefish VM with MMTk!