Fix nasty bug with `ChunkMap` when using two or more spaces that access it
`ChunkMap` was a policy-local metadata but it was encoding global state
(which chunk has been allocated). If you use two or more policies that
use the `ChunkMap` at the same time (for example an `ImmixSpace` and a
`MarkSweepSpace`) then you may end up unsetting the chunk-allocated
state resulting in bugs down the line.

This commit fixes this by encoding the chunk owner into the remaining
bits we have for the metadata and by lifting the chunk map metadata to
be a global metadata instead of a local one.