Allow Immix to work with optimized single-threaded build
This commit allows us to run Immix with the optimized single-threaded
build. This now allows us to run the single-threaded build with an APEX
build as well.

This commit fixes a bug with chunk iteration as well so that we don't
accidentally operate on a chunk that does not belong to us.