Don't count mutator transition states towards the mutator
With this change, these transition states are being counted towards the
STW time. This is also not ideal, but it's a hack for the time-being.
The runtime/binding also needs to set the GC status to
`GcState::NotInGC` correctly right after mutators have started.

See https://github.com/mmtk/mmtk-core/issues/1330 for more information.